Consolidate client knobs header files
