+++ fdbserver/ratekeeper/Ratekeeper.cpp
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-#include "fdbclient/ClientKnobs.h"
+#include "fdbclient/Knobs.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/ServerDBInfo.h"
 #include "fdbserver/core/WaitFailure.h"
